pyproject.toml
[project]
name = "drawio-to-pptx"
version = "0.1.0"
description = "Drawio XML to .drawio and editable PPTX pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-pptx>=0.6.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
drawio-to-pptx = "drawio_to_pptx.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["drawio_to_pptx"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
